requirements.txt
numpy==2.3.5
pandas==2.3.3
psycopg==3.2.12
psycopg-binary==3.2.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.2